demo fix bug

+++ YNSelfSizingCollectionViewDemo/YNDemoFixedWidthAndHeightCollectionViewCell.m
@@ -47,7 +47,7 @@
 }
 
 -(CGFloat)fixedWidth{
-    return self.collectionView.bounds.size.width / 4.f;
+    return (int)(self.collectionView.bounds.size.width / 4.f);
 }
 
 -(CGFloat)fixedHeight{
